extracted disruption evaluation logic to separate class
